Adds a put call to the Device.unlock method.
Testing indicates the unlock method's test was passing because of problems with the datastore_v3 stub, which contained an updated device entity even in the absence of a Device.put call. This change separates the lock and unlock unit tests, and skips the Device.enroll method to use an in-code Device entity. (I confirmed that in the absence of a put call in Device.unlock, the unlock test now fails.)

PiperOrigin-RevId: 195860029