Changes Action behavior to prevent datastore collisions.
Async Actions, which run as tasks, can overwrite each others' changes on a given entity because we don't tightly control their execution. This CL causes async Actions to run sequentially, with each task spawning the next, allowing us to wait for datastore puts on an entity in one task before getting the entity for further modification in a subsequent task. The new flow is the following:

1. When part of the code calls raise_event for a datastore model, the event lib runs all sync actions sequentially, as before.
2. When the method moves on to the async actions, instead of spawning tasks for each of them it passes an alphabetized list of them to a single task as a kwarg, async_actions, which replaces the old action_name kwarg.
3. The first task pops item 0 from the list and performs that named action. If there are still items in the list, it spawns a new task with the same payload (though the async_actions list is now one item shorter).

Because this creates a chain that could be interrupted by an exception, we use broad exception handling around the call of the action's run method to ensure further tasks will run.

PiperOrigin-RevId: 209974299