Adding get() method to the Tag API and generalizing tag messages.
This CL renames the get_by_urlsafe_key() method in the model to get() because get_by_urlsafe_key() was named to disambiguate from get_by_name(), which we implemented in a pre_put_hook instead. This also adds a get() API method to match the model, which will be used when viewing tag details on the FE. The DestroyTagRequest was renamed TagRequest so it can be used for any request where only a urlsafe key is necessary such as destroy, get, and future methods.

PiperOrigin-RevId: 225473617